move linux-tools installation from ppc to pc-sanity temporarily
for https://code.launchpad.net/~alextu/plainbox-provider-checkbox/+git/plainbox-provider-checkbox/+merge/388726
It's not accepted to install packages in test cases.
So, we need to install linux-tools before testing.
The final goal would be install the packages before checkbox testing.
e.g. prepare-checkbox-sanity or testflinger yaml.

This change is to make sure linux-tools be installed before we achived the final
goal.