Improve miscellanea/check-nvidia (LP: #1980173)
As the check_nvidia.sh is assuming the nvidia card is installed. If any
graphic card w/o driver, then it will show 'pcieport' which causes
check_nvidia.sh be executed. As check_nvidia.sh is specify on propritery
nvidia driver. Make it only be executed if 'nvidia' there to reduce the
noise. (LP: #1980173)